Split test_comparator.py into per-module test files
- comparator/test_utils.py: utils function tests
- comparator/tensor_comparison/test_compare.py: compare_tensors,
  _compute_diff, _compute_tensor_stats tests
- comparator/tensor_comparison/test_printer.py: print_comparison tests
- comparator/test_entrypoint.py: E2E test with main()
